pyproject.toml
[project]
name = "smart-home-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.3",
    "tapo>=0.8.11",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
